server/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart
python-dotenv
httpx
aiofiles
asyncio
aiohttp
tenacity
psutil
torch
transformers
accelerate
diffusers
safetensors
Pillow
opencv-python
imageio
numpy
pandas
scikit-learn==1.6.1
joblib
xgboost
requests
web3
huggingface_hub
google-generativeai
imageio[ffmpeg]
av